Changed Structure, Added Features
Changed the structure of the bot to have more functions instead of lines of code inside main. Also changed it to where the daily tweet includes the closest graduation and a random future one as opposed to the second closest.

The more important feature added is the ability to tweet at the bot a season and year and it will respond with the amount of days until your graduation. For example:
@Bot Winter 2017
will return
You have xx days until your graduation!